refactor: rename settings to setting
