+++ src/app/i18n/helpers/index.ts
@@ -6,7 +6,7 @@ import { translations as OFFLINE } from "../../../domains/offline/i18n";
 import { translations as PLUGINS } from "../../../domains/plugins/i18n";
 import { translations as PROFILE } from "../../../domains/profile/i18n";
 import { translations as SEARCH } from "../../../domains/search/i18n";
-import { translations as SETTINGS } from "../../../domains/settings/i18n";
+import { translations as SETTINGS } from "../../../domains/setting/i18n";
 import { translations as SUPPORT } from "../../../domains/support/i18n";
 import { translations as TRANSACTION } from "../../../domains/transaction/i18n";
 import { translations as WALLETS } from "../../../domains/wallets/i18n";
