+++ src/domains/setting/i18n.ts
