+++ src/domains/setting/pages/Settings/available-settings/General.tsx
